Fix various warnings on Linux, now virtually warning-free.
Add new bigram probability table to pregen

git-svn-id: https://groke.otago.ac.nz/ant/ant@1294 7489a7b4-4588-4892-8acd-774bacfab51c

+++ tools/make_unicode_tables.c
@@ -367,7 +367,7 @@ for (current = lines; *current != NULL; current++)
 			}
 		*utf8_buf_ptr = 0;
 
-		longest_decomposition = max(longest_decomposition, utf8_buf_ptr - utf8_buf);
+		longest_decomposition = max(longest_decomposition, (int) (utf8_buf_ptr - utf8_buf));
 
 		printf("{%ld, \"%s\"}", character, utf8_buf);
 		times++;
